refactor: split StatusDashboard component
